refactor(blocks): organize code executor into submodules
