--- autogpt_platform/backend/backend/blocks/code_executor.py
@@ -1,676 +1,0 @@
-from typing import TYPE_CHECKING, Any, Literal, Optional
-
-from e2b_code_interpreter import AsyncSandbox
-from e2b_code_interpreter import Result as E2BExecutionResult
-from e2b_code_interpreter.charts import Chart as E2BExecutionResultChart
-from pydantic import BaseModel, Field, JsonValue, SecretStr
-
-from backend.blocks._base import (
-    Block,
-    BlockCategory,
-    BlockOutput,
-    BlockSchemaInput,
-    BlockSchemaOutput,
-)
-from backend.blocks.code_executor_desktop import (
-    create_desktop_sandbox,
-    kill_desktop_sandbox,
-)
-from backend.blocks.code_executor_helpers import (
-    ProgrammingLanguage,
-    build_variable_injection,
-)
-from backend.data.model import (
-    APIKeyCredentials,
-    CredentialsField,
-    CredentialsMetaInput,
-    SchemaField,
-)
-from backend.integrations.providers import ProviderName
-from backend.util.desktop_preview import create_preview_link
-from backend.util.sandbox_files import (
-    SandboxFileOutput,
-    extract_and_store_sandbox_files,
-)
-
-if TYPE_CHECKING:
-    from backend.executor.utils import ExecutionContext
-
-TEST_CREDENTIALS = APIKeyCredentials(
-    id="01234567-89ab-cdef-0123-456789abcdef",
-    provider="e2b",
-    api_key=SecretStr("mock-e2b-api-key"),
-    title="Mock E2B API key",
-    expires_at=None,
-)
-
-
-def _mock_execute_code(
-    *args: Any, **kwargs: Any
-) -> tuple[list, str, str, str, str, list]:
-    """Stub for `execute_code` used by the blocks' test_mock.
-
-    Returns the (results, text, stdout, stderr, sandbox_id, files) tuple,
-    echoing back any provided `sandbox_id` so the step block's test sees it.
-    """
-    sandbox_id = kwargs.get("sandbox_id") or "sandbox_id"
-    return [], "Hello World", "Hello World\n", "", sandbox_id, []
-
-
-TEST_CREDENTIALS_INPUT = {
-    "provider": TEST_CREDENTIALS.provider,
-    "id": TEST_CREDENTIALS.id,
-    "type": TEST_CREDENTIALS.type,
-    "title": TEST_CREDENTIALS.type,
-}
-
-
-_MAIN_RESULT_DESCRIPTION = (
-    "The main result from the code execution (the script's final "
-    "expression). Its `json` sub-field is ONLY populated when the "
-    "result is a dict/object/map — bare lists, strings, and "
-    "numbers land in `text` as a string instead. To pass "
-    "structured data downstream via `main_result_#_json_#_<key>` "
-    "links, end the script with a key-value structure in the "
-    "script's language (e.g. `{'items': my_list}` in Python, "
-    "`({items: myList})` in JavaScript)."
-)
-
-
-class MainCodeExecutionResult(BaseModel):
-    """
-    *Pydantic model mirroring `e2b_code_interpreter.Result`*
-
-    Represents the data to be displayed as a result of executing a cell in a Jupyter notebook.
-    The result is similar to the structure returned by ipython kernel: https://ipython.readthedocs.io/en/stable/development/execution.html#execution-semantics
-
-    The result can contain multiple types of data, such as text, images, plots, etc. Each type of data is represented
-    as a string, and the result can contain multiple types of data. The display calls don't have to have text representation,
-    for the actual result the representation is always present for the result, the other representations are always optional.
-    """  # noqa
-
-    class Chart(BaseModel, E2BExecutionResultChart):
-        pass
-
-    text: Optional[str] = None
-    html: Optional[str] = None
-    markdown: Optional[str] = None
-    svg: Optional[str] = None
-    png: Optional[str] = None
-    jpeg: Optional[str] = None
-    pdf: Optional[str] = None
-    latex: Optional[str] = None
-    json_data: Optional[JsonValue] = Field(None, alias="json")
-    javascript: Optional[str] = None
-    data: Optional[dict] = None
-    chart: Optional[Chart] = None
-    extra: Optional[dict] = None
-    """Extra data that can be included. Not part of the standard types."""
-
-
-class CodeExecutionResult(MainCodeExecutionResult):
-    __doc__ = MainCodeExecutionResult.__doc__
-
-    is_main_result: bool = False
-    """Whether this data is the main result of the cell. Data can be produced by display calls of which can be multiple in a cell."""  # noqa
-
-
-class BaseE2BExecutorMixin:
-    """Shared implementation methods for E2B executor blocks."""
-
-    # Default working directory in E2B sandboxes
-    WORKING_DIR = "/home/user"
-
-    async def execute_code(
-        self,
-        api_key: str,
-        code: str,
-        language: ProgrammingLanguage,
-        template_id: str = "",
-        setup_commands: Optional[list[str]] = None,
-        timeout: Optional[int] = None,
-        sandbox_id: Optional[str] = None,
-        dispose_sandbox: bool = False,
-        execution_context: Optional["ExecutionContext"] = None,
-        extract_files: bool = False,
-        envs: Optional[dict[str, str]] = None,
-    ):
-        """
-        Create or reconnect to a sandbox, then run setup commands and code.
-
-        Setup commands run on every invocation, including reconnects, so callers
-        reusing them must make them idempotent.
-
-        Args:
-            extract_files: If True and execution_context provided, extract files
-                           created/modified during execution and store to workspace.
-        """  # noqa
-        sandbox = None
-        files: list[SandboxFileOutput] = []
-        try:
-            if sandbox_id:
-                sandbox = await AsyncSandbox.connect(
-                    sandbox_id=sandbox_id, api_key=api_key, timeout=timeout
-                )
-            else:
-                sandbox = await AsyncSandbox.create(
-                    api_key=api_key, template=template_id, timeout=timeout
-                )
-            if setup_commands:
-                for cmd in setup_commands:
-                    await sandbox.commands.run(cmd)
-
-            # Capture timestamp before execution to scope file extraction
-            start_timestamp = None
-            if extract_files:
-                ts_result = await sandbox.commands.run("date -u +%Y-%m-%dT%H:%M:%S")
-                start_timestamp = ts_result.stdout.strip() if ts_result.stdout else None
-
-            # Execute the code
-            execution = await sandbox.run_code(  # type: ignore[attr-defined]
-                code,
-                language=language.value,
-                envs=envs or {},
-                on_error=lambda e: sandbox.kill(),  # Kill the sandbox on error
-            )
-
-            if execution.error:
-                raise Exception(execution.error)
-
-            results = execution.results
-            text_output = execution.text
-            stdout_logs = "".join(execution.logs.stdout)
-            stderr_logs = "".join(execution.logs.stderr)
-
-            # Extract files created/modified during this execution
-            if extract_files and execution_context:
-                files = await extract_and_store_sandbox_files(
-                    sandbox=sandbox,
-                    working_directory=self.WORKING_DIR,
-                    execution_context=execution_context,
-                    since_timestamp=start_timestamp,
-                    text_only=False,  # Include binary files too
-                )
-
-            return (
-                results,
-                text_output,
-                stdout_logs,
-                stderr_logs,
-                sandbox.sandbox_id,
-                files,
-            )
-        finally:
-            # Dispose of sandbox if requested to reduce usage costs
-            if dispose_sandbox and sandbox:
-                await sandbox.kill()
-
-    def process_execution_results(
-        self, results: list[E2BExecutionResult]
-    ) -> tuple[dict[str, Any] | None, list[dict[str, Any]]]:
-        """Process and filter execution results."""
-        # Filter out empty formats and convert to dicts
-        processed_results = [
-            {
-                f: value
-                for f in [*r.formats(), "extra", "is_main_result"]
-                if (value := getattr(r, f, None)) is not None
-            }
-            for r in results
-        ]
-        if main_result := next(
-            (r for r in processed_results if r.get("is_main_result")), None
-        ):
-            # Make main_result a copy we can modify & remove is_main_result
-            (main_result := {**main_result}).pop("is_main_result")
-
-        return main_result, processed_results
-
-
-class ExecuteCodeBlock(Block, BaseE2BExecutorMixin):
-    # TODO : Add support to upload and download files
-    # NOTE: Currently, you can only customize the CPU and Memory
-    #       by creating a pre customized sandbox template
-    class Input(BlockSchemaInput):
-        credentials: CredentialsMetaInput[
-            Literal[ProviderName.E2B], Literal["api_key"]
-        ] = CredentialsField(
-            description=(
-                "Enter your API key for the E2B platform. "
-                "You can get it in here - https://e2b.dev/docs"
-            ),
-        )
-
-        # Todo : Option to run commond in background
-        setup_commands: list[str] = SchemaField(
-            description=(
-                "Shell commands to set up the sandbox before running the code. "
-                "You can use `curl` or `git` to install your desired Debian based "
-                "package manager. `pip` and `npm` are pre-installed.\n\n"
-                "These commands are executed with `sh`, in the foreground."
-            ),
-            placeholder="pip install cowsay",
-            default_factory=list,
-            advanced=False,
-        )
-
-        variables: dict[str, Any] = SchemaField(
-            title="Variables (Python/JS only)",
-            description=(
-                "Variables defined here can be used directly in your code. "
-                "Each key (`variables_#_{name}`) is injected directly as a local "
-                "variable with the same name (`{name}`) in your code. "
-                "Values wired in from other blocks keep their type; default values set "
-                "on this node come in as strings, so parse them in your code "
-                "if you need a number or other type."
-            ),
-            default_factory=dict,
-            advanced=False,
-        )
-
-        code: str = SchemaField(
-            description="Code to execute in the sandbox",
-            placeholder="print('Hello, World!')",
-            default="",
-            advanced=False,
-        )
-
-        language: ProgrammingLanguage = SchemaField(
-            description="Programming language to execute",
-            default=ProgrammingLanguage.PYTHON,
-            advanced=False,
-        )
-
-        timeout: int = SchemaField(
-            description="Sandbox lifetime in seconds from creation", default=300
-        )
-
-        dispose_sandbox: bool = SchemaField(
-            description=(
-                "Whether to dispose of the sandbox immediately after execution. "
-                "If disabled, the sandbox will run until its timeout expires."
-            ),
-            default=True,
-        )
-
-        template_id: str = SchemaField(
-            description=(
-                "You can use an E2B sandbox template by entering its ID here. "
-                "Check out the E2B docs for more details: "
-                "[E2B - Sandbox template](https://e2b.dev/docs/sandbox-template)"
-            ),
-            default="",
-            advanced=True,
-        )
-
-    class Output(BlockSchemaOutput):
-        main_result: MainCodeExecutionResult = SchemaField(
-            title="Main Result",
-            description=_MAIN_RESULT_DESCRIPTION,
-        )
-        results: list[CodeExecutionResult] = SchemaField(
-            description="List of results from the code execution"
-        )
-        response: str = SchemaField(
-            title="Main Text Output",
-            description="Text output (if any) of the main execution result",
-        )
-        stdout_logs: str = SchemaField(
-            description="Standard output logs from execution"
-        )
-        stderr_logs: str = SchemaField(description="Standard error logs from execution")
-        files: list[SandboxFileOutput] = SchemaField(
-            description=(
-                "Files created or modified during execution. "
-                "Each file has path, name, content, and workspace_ref (if stored)."
-            ),
-        )
-
-    def __init__(self):
-        super().__init__(
-            id="0b02b072-abe7-11ef-8372-fb5d162dd712",
-            description="Executes code in a sandbox environment with internet access.",
-            categories={BlockCategory.DEVELOPER_TOOLS},
-            input_schema=ExecuteCodeBlock.Input,
-            output_schema=ExecuteCodeBlock.Output,
-            test_credentials=TEST_CREDENTIALS,
-            test_input={
-                "credentials": TEST_CREDENTIALS_INPUT,
-                "code": "print('Hello World')",
-                "language": ProgrammingLanguage.PYTHON.value,
-                "setup_commands": [],
-                "timeout": 300,
-                "template_id": "",
-            },
-            test_output=[
-                ("results", []),
-                ("response", "Hello World"),
-                ("stdout_logs", "Hello World\n"),
-                ("files", []),
-            ],
-            test_mock={"execute_code": _mock_execute_code},
-        )
-
-    async def run(
-        self,
-        input_data: Input,
-        *,
-        credentials: APIKeyCredentials,
-        execution_context: "ExecutionContext",
-        **kwargs,
-    ) -> BlockOutput:
-        try:
-            # Expose user-provided variables by passing them as a JSON env var and
-            # prepending a constant snippet that deserializes them into the runtime.
-            # Keeping the data in the env var (not the code string) avoids injection.
-            envs, prefix = build_variable_injection(
-                input_data.variables, input_data.language
-            )
-            code = prefix + input_data.code
-
-            results, text_output, stdout, stderr, _, files = await self.execute_code(
-                api_key=credentials.api_key.get_secret_value(),
-                code=code,
-                language=input_data.language,
-                template_id=input_data.template_id,
-                setup_commands=input_data.setup_commands,
-                timeout=input_data.timeout,
-                dispose_sandbox=input_data.dispose_sandbox,
-                execution_context=execution_context,
-                extract_files=True,
-                envs=envs,
-            )
-
-            # Determine result object shape & filter out empty formats
-            main_result, results = self.process_execution_results(results)
-            if main_result:
-                yield "main_result", main_result
-            yield "results", results
-            if text_output:
-                yield "response", text_output
-            if stdout:
-                yield "stdout_logs", stdout
-            if stderr:
-                yield "stderr_logs", stderr
-            # Always yield files (empty list if none)
-            yield "files", [f.model_dump() for f in files]
-        except Exception as e:
-            yield "error", str(e)
-
-
-class InstantiateCodeSandboxBlock(Block, BaseE2BExecutorMixin):
-    class Input(BlockSchemaInput):
-        credentials: CredentialsMetaInput[
-            Literal[ProviderName.E2B], Literal["api_key"]
-        ] = CredentialsField(
-            description=(
-                "Enter your API key for the E2B platform. "
-                "You can get it in here - https://e2b.dev/docs"
-            )
-        )
-
-        # Todo : Option to run commond in background
-        setup_commands: list[str] = SchemaField(
-            description=(
-                "Shell commands to set up the sandbox before running the code. "
-                "You can use `curl` or `git` to install your desired Debian based "
-                "package manager. `pip` and `npm` are pre-installed.\n\n"
-                "These commands are executed with `sh`, in the foreground."
-            ),
-            placeholder="pip install cowsay",
-            default_factory=list,
-            advanced=False,
-        )
-
-        setup_code: str = SchemaField(
-            description="Code to execute in the sandbox",
-            placeholder="print('Hello, World!')",
-            default="",
-            advanced=False,
-        )
-
-        language: ProgrammingLanguage = SchemaField(
-            description="Programming language to execute",
-            default=ProgrammingLanguage.PYTHON,
-            advanced=False,
-        )
-
-        timeout: int = SchemaField(
-            description=(
-                "Sandbox lifetime in seconds. Choose enough time to run setup and "
-                "test through the live URL (up to 3600 seconds)."
-            ),
-            default=300,
-            ge=1,
-            le=3600,
-        )
-
-        enable_live_view: bool = SchemaField(
-            description=(
-                "Start an interactive desktop preview and return live_url. Requires a custom "
-                "template_id containing both the desktop and code interpreter. "
-                "Use the returned sandbox_id with Execute Code Step to work in "
-                "the same environment. The preview ends when the sandbox stops."
-            ),
-            default=False,
-            advanced=True,
-        )
-
-        template_id: str = SchemaField(
-            description=(
-                "You can use an E2B sandbox template by entering its ID here. "
-                "Check out the E2B docs for more details: "
-                "[E2B - Sandbox template](https://e2b.dev/docs/sandbox-template)"
-            ),
-            default="",
-            advanced=True,
-        )
-
-    class Output(BlockSchemaOutput):
-        sandbox_id: str = SchemaField(description="ID of the sandbox instance")
-        live_url: str = SchemaField(
-            description=(
-                "Desktop preview link for the signed-in user who created the sandbox. "
-                "Returned after setup when live view is enabled. The link expires after 24 hours."
-            )
-        )
-        response: str = SchemaField(
-            title="Text Result",
-            description="Text result (if any) of the setup code execution",
-        )
-        stdout_logs: str = SchemaField(
-            description="Standard output logs from execution"
-        )
-        stderr_logs: str = SchemaField(description="Standard error logs from execution")
-
-    def __init__(self):
-        super().__init__(
-            id="ff0861c9-1726-4aec-9e5b-bf53f3622112",
-            description=(
-                "Instantiate a sandbox environment with internet access "
-                "in which you can execute code with the Execute Code Step block. "
-                "Optionally enable a live desktop preview and return its viewing URL."
-            ),
-            categories={BlockCategory.DEVELOPER_TOOLS},
-            input_schema=InstantiateCodeSandboxBlock.Input,
-            output_schema=InstantiateCodeSandboxBlock.Output,
-            test_credentials=TEST_CREDENTIALS,
-            test_input={
-                "credentials": TEST_CREDENTIALS_INPUT,
-                "setup_code": "print('Hello World')",
-                "language": ProgrammingLanguage.PYTHON.value,
-                "setup_commands": [],
-                "timeout": 300,
-                "template_id": "",
-            },
-            test_output=[
-                ("sandbox_id", str),
-                ("response", "Hello World"),
-                ("stdout_logs", "Hello World\n"),
-            ],
-            test_mock={"execute_code": _mock_execute_code},
-        )
-
-    async def run(
-        self,
-        input_data: Input,
-        *,
-        credentials: APIKeyCredentials,
-        user_id: str = "",
-        **kwargs,
-    ) -> BlockOutput:
-        desktop_id = None
-        sandbox_handed_off = False
-        live_url = None
-        api_key = credentials.api_key.get_secret_value()
-        try:
-            if input_data.enable_live_view:
-                if not user_id:
-                    raise ValueError("Live view requires an authenticated user")
-                desktop_id, live_url = await create_desktop_sandbox(
-                    api_key=api_key,
-                    template_id=input_data.template_id,
-                    timeout=input_data.timeout,
-                )
-
-            _, text_output, stdout, stderr, sandbox_id, _ = await self.execute_code(
-                api_key=api_key,
-                code=input_data.setup_code,
-                language=input_data.language,
-                template_id=input_data.template_id,
-                setup_commands=input_data.setup_commands,
-                timeout=input_data.timeout,
-                sandbox_id=desktop_id,
-            )
-            if not sandbox_id:
-                yield "error", "Sandbox ID not found"
-                return
-            if live_url:
-                yield "live_url", create_preview_link(user_id, live_url)
-            sandbox_handed_off = True
-            yield "sandbox_id", sandbox_id
-
-            if text_output:
-                yield "response", text_output
-            if stdout:
-                yield "stdout_logs", stdout
-            if stderr:
-                yield "stderr_logs", stderr
-        except Exception as e:
-            yield "error", str(e)
-        finally:
-            if desktop_id and not sandbox_handed_off:
-                await kill_desktop_sandbox(api_key, desktop_id)
-
-
-class ExecuteCodeStepBlock(Block, BaseE2BExecutorMixin):
-    class Input(BlockSchemaInput):
-        credentials: CredentialsMetaInput[
-            Literal[ProviderName.E2B], Literal["api_key"]
-        ] = CredentialsField(
-            description=(
-                "Enter your API key for the E2B platform. "
-                "You can get it in here - https://e2b.dev/docs"
-            ),
-        )
-
-        sandbox_id: str = SchemaField(
-            description="ID of the sandbox instance to execute the code in",
-            advanced=False,
-        )
-
-        step_code: str = SchemaField(
-            description="Code to execute in the sandbox",
-            placeholder="print('Hello, World!')",
-            default="",
-            advanced=False,
-        )
-
-        language: ProgrammingLanguage = SchemaField(
-            description="Programming language to execute",
-            default=ProgrammingLanguage.PYTHON,
-            advanced=False,
-        )
-
-        dispose_sandbox: bool = SchemaField(
-            description="Whether to dispose of the sandbox after executing this code.",
-            default=False,
-        )
-
-        timeout: Optional[int] = SchemaField(
-            description=(
-                "Extend the remaining sandbox lifetime to at least this many seconds "
-                "when connecting (up to 3600). Cannot shorten a longer remaining lifetime. "
-                "If omitted, E2B's default extension applies. Use dispose_sandbox to stop early."
-            ),
-            default=None,
-            ge=1,
-            le=3600,
-            advanced=False,
-        )
-
-    class Output(BlockSchemaOutput):
-        main_result: MainCodeExecutionResult = SchemaField(
-            title="Main Result",
-            description=_MAIN_RESULT_DESCRIPTION,
-        )
-        results: list[CodeExecutionResult] = SchemaField(
-            description="List of results from the code execution"
-        )
-        response: str = SchemaField(
-            title="Main Text Output",
-            description="Text output (if any) of the main execution result",
-        )
-        stdout_logs: str = SchemaField(
-            description="Standard output logs from execution"
-        )
-        stderr_logs: str = SchemaField(description="Standard error logs from execution")
-
-    def __init__(self):
-        super().__init__(
-            id="82b59b8e-ea10-4d57-9161-8b169b0adba6",
-            description="Execute code in a previously instantiated sandbox.",
-            categories={BlockCategory.DEVELOPER_TOOLS},
-            input_schema=ExecuteCodeStepBlock.Input,
-            output_schema=ExecuteCodeStepBlock.Output,
-            test_credentials=TEST_CREDENTIALS,
-            test_input={
-                "credentials": TEST_CREDENTIALS_INPUT,
-                "sandbox_id": "sandbox_id",
-                "step_code": "print('Hello World')",
-                "language": ProgrammingLanguage.PYTHON.value,
-            },
-            test_output=[
-                ("results", []),
-                ("response", "Hello World"),
-                ("stdout_logs", "Hello World\n"),
-            ],
-            test_mock={"execute_code": _mock_execute_code},
-        )
-
-    async def run(
-        self, input_data: Input, *, credentials: APIKeyCredentials, **kwargs
-    ) -> BlockOutput:
-        try:
-            results, text_output, stdout, stderr, _, _ = await self.execute_code(
-                api_key=credentials.api_key.get_secret_value(),
-                code=input_data.step_code,
-                language=input_data.language,
-                sandbox_id=input_data.sandbox_id,
-                dispose_sandbox=input_data.dispose_sandbox,
-                timeout=input_data.timeout,
-            )
-
-            # Determine result object shape & filter out empty formats
-            main_result, results = self.process_execution_results(results)
-            if main_result:
-                yield "main_result", main_result
-            yield "results", results
-            if text_output:
-                yield "response", text_output
-            if stdout:
-                yield "stdout_logs", stdout
-            if stderr:
-                yield "stderr_logs", stderr
-        except Exception as e:
-            yield "error", str(e)
