+++ autogpt_platform/backend/backend/blocks/code_executor/desktop.py
