+++ autogpt_platform/backend/backend/blocks/code_executor/desktop_cleanup_test.py
@@ -5,12 +5,9 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from backend.blocks.code_executor import (
-    TEST_CREDENTIALS,
-    TEST_CREDENTIALS_INPUT,
-    InstantiateCodeSandboxBlock,
-)
-from backend.blocks.code_executor_desktop import kill_desktop_sandbox
+from backend.blocks.code_executor._test import TEST_CREDENTIALS, TEST_CREDENTIALS_INPUT
+from backend.blocks.code_executor.desktop import kill_desktop_sandbox
+from backend.blocks.code_executor.instantiate import InstantiateCodeSandboxBlock
 
 
 async def test_cancelled_cleanup_continues_until_kill_finishes():
@@ -21,7 +18,7 @@ async def test_cancelled_cleanup_continues_until_kill_finishes():
         await finish.wait()
         killed.set()
 
-    with patch("backend.blocks.code_executor_desktop.AsyncSandbox.kill", kill):
+    with patch("backend.blocks.code_executor.desktop.AsyncSandbox.kill", kill):
         task = asyncio.create_task(kill_desktop_sandbox("key", "desktop-id"))
         await started.wait()
         task.cancel()
@@ -42,8 +39,8 @@ async def test_hung_cleanup_times_out_and_logs_reason(caplog):
             stopped.set()
 
     with (
-        patch("backend.blocks.code_executor_desktop.AsyncSandbox.kill", kill),
-        patch("backend.blocks.code_executor_desktop.CLEANUP_TIMEOUT", 0.01),
+        patch("backend.blocks.code_executor.desktop.AsyncSandbox.kill", kill),
+        patch("backend.blocks.code_executor.desktop.CLEANUP_TIMEOUT", 0.01),
     ):
         await asyncio.wait_for(kill_desktop_sandbox("key", "desktop-id"), timeout=1)
     assert stopped.is_set()
@@ -58,7 +55,7 @@ async def test_missing_setup_id_never_hands_off_desktop():
     )
     with (
         patch(
-            "backend.blocks.code_executor.create_desktop_sandbox",
+            "backend.blocks.code_executor.instantiate.create_desktop_sandbox",
             AsyncMock(return_value=("desktop-id", "https://preview.example")),
         ),
         patch.object(
@@ -66,7 +63,9 @@ async def test_missing_setup_id_never_hands_off_desktop():
             "execute_code",
             AsyncMock(return_value=([], "partial", "partial", "partial", "", [])),
         ),
-        patch("backend.blocks.code_executor.kill_desktop_sandbox", AsyncMock()) as kill,
+        patch(
+            "backend.blocks.code_executor.instantiate.kill_desktop_sandbox", AsyncMock()
+        ) as kill,
     ):
         outputs = dict(
             [
