+++ autogpt_platform/backend/backend/blocks/code_executor/helpers.py
