+++ docs/integrations/SUMMARY.md
@@ -44,6 +44,9 @@
 * [Baas Bots](block-integrations/baas/bots.md)
 * [Bannerbear Text Overlay](block-integrations/bannerbear/text_overlay.md)
 * [Basic](block-integrations/basic.md)
+* [Code Executor Execute](block-integrations/code_executor/execute.md)
+* [Code Executor Instantiate](block-integrations/code_executor/instantiate.md)
+* [Code Executor Step](block-integrations/code_executor/step.md)
 * [Compass Triggers](block-integrations/compass/triggers.md)
 * [Data](block-integrations/data.md)
 * [DataForB2B Enrich](block-integrations/dataforb2b/enrich.md)
